requirements.txt
torch ~= 2.1.0
pytorch-lightning ~= 2.1.3
pandas ~= 2.1.4
huggingface-hub ~= 0.20.1
dvs ~= 3.38.1
dvc[gdrive]~=3.0.1
# The next three I had to install manually to get `datasets` to work.
multiprocess ~= 0.70.15
xxhash ~= 3.4.1
pyarrow ~= 14.0.2
datasets ~= 2.16.1
hydra-core ~= 1.3.2
transformers ~= 4.36.2
wandb ~= 0.16.1
fastapi ~= 0.108.0
uvicorn[standard] ~= 0.25.0
